consolidated data types to 1 file, updated imports
