+++ src/App.tsx
@@ -4,7 +4,7 @@ import "./App.css";
 import WorldMaps from "./pages/WorldMaps";
 import NoPage from "./pages/NoPage";
 import Loading from "./pages/Loading";
-import { WorldMapData } from "./types/worldMapTypes";
+import { WorldMapData } from "./types/dataTypes";
 import { MapData } from "./types/mapTypes";
 import { MobData } from "./types/mobTypes";
 import { useFetchMapIds } from "./hooks/useFetchMapIds";
