--- src/types/worldMapTypes.ts
@@ -1,22 +1,0 @@
-export type WorldMapData = {
-  worldMapName?: string;
-  parentWorld?: string;
-  links: Link[];
-  maps: Map[];
-};
-
-export type Link = {
-  linksTo: string;
-  imageBase64: string;
-  x: number;
-  y: number;
-};
-
-export type Map = {
-  noTooltip?: boolean;
-  description?: string;
-  type: number;
-  x: number;
-  y: number;
-  mapNumbers: number[];
-};
